chore: removed unused types and moved files
