+++ package.json
@@ -51,7 +51,7 @@
     "node-cron": "^3.0.3",
     "pino": "^8.15.0",
     "pino-pretty": "^10.2.0",
-    "skyhelper-utils": "^1.2.16",
+    "skyhelper-utils": "^1.2.19",
     "table": "^6.8.1",
     "topgg-autoposter": "^2.0.2"
   },
@@ -88,6 +88,9 @@
     ],
     "#api/*": [
       "./dist/api/*.js"
+    ],
+    "#utils": [
+      "./dist/bot/utils/index.js"
     ]
   },
   "devDependencies": {
