+++ src/bot/handlers/index.ts
@@ -1,8 +1,5 @@
 export { default as logger } from "#handlers/logger";
-export { default as buildShardEmbed } from "#handlers/buildShardsEmbed";
-export * from "#bot/handlers/buildEventsEmbed";
 export { default as getTS } from "#bot/utils/getTSStatus";
 export { default as dblStats } from "#handlers/dblStats";
 export { default as topggAutopost } from "#handlers/topggAutopost";
 export * from "#handlers/sendDailyQuestReminder";
-export * from "#handlers/dailyQuestEmbed";
