+++ src/bot/libs/constants/index.ts
@@ -2,4 +2,3 @@ export { default as spiritsData } from "./spirits-datas/index.js";
 export * from "#libs/constants/seasonPrices";
 export { default as seasonsData } from "#libs/constants/seasonsData";
 export * from "./realmsData.js";
-export * from "./eventData.js";
