+++ src/bot/utils/buildEventsEmbed.ts
