+++ src/bot/utils/buildShardsEmbed.ts
