+++ src/bot/utils/dailyQuestEmbed.ts
