+++ src/bot/utils/index.ts
@@ -1,1 +1,4 @@
 export { default as getTSStatus } from "./getTSStatus.js";
+export * from "./buildEventsEmbed.js";
+export { default as buildShardEmbed } from "./buildShardsEmbed.js";
+export * from "./dailyQuestEmbed.js";
